add clear all receipts delete function

+++ src/main/java/models/Item.java
@@ -51,7 +51,7 @@ public class Item {
         return this.id;
     }
 
-    public double getSplit() {
+    public int getSplit() {
        return this.split;
     }
 }
